test(core): split watch lifecycle tests
