+++ packages/core/test/core/watch/basic-lifecycle/helpers.ts
@@ -1,0 +1,14 @@
+export {
+  describe,
+  test,
+  expect,
+  watch,
+  vitePlugin,
+  mkdtempSync,
+  writeFileSync,
+  rmSync,
+  join,
+  tmpdir,
+} from '../helpers';
+
+export type { RollupPlugin, ZntcPlugin } from '../helpers';
